Put file upload in a new FileUpload component (in #669)
